refactor and change example to tranaction records, because better for debugging
