+++ src/routes/user/transaction_records/ProtectedRouteUserTransactionRecord.js
@@ -1,8 +1,8 @@
 import React, { useContext } from 'react';
 import { Navigate } from 'react-router-dom';
-import { AuthContext } from '../../login/AuthContext';
-import { LanguageContext } from '../../context/LanguageContext';
-import UserTransactionRecords from '../../data-components/user/user-transaction-record';  // Schreibfehler korrigiert
+import { AuthContext } from '../../../login/AuthContext';
+import { LanguageContext } from '../../../context/LanguageContext';
+import UserTransactionRecords from '../../../data-components/user/user-transaction-record/user-transaction-record';  // Schreibfehler korrigiert
 
 const translations = {
     de: {
